"Removed unnecessary whitespace in `monitor.go`
The commit removes additional whitespace which was present in the `monitor.go` file within the `premint` module. The removal of the whitespace doesn't affect the functionality of the code but it makes the code cleaner and easier to read, following conventional formatting standards. This change will be helpful in maintaining the readability and neatness of the codebase."