Add new modules and refactor webhook handling
Extended the supported module list in types.go by adding Etherscan, ExchangeArt, LaunchMyNFT, OpenSea, and Ethereum. Also included an EtherscanWebhook in the WebhookConfig struct.
In webhook.go, generalised the webhook handling by creating a 'Handler' function that takes a webhook and a module, marshalling the webhook content and pushing it to the appropriate module's webhook url. It reduces repetitions and increase the code's maintainability. Undo the change will require reverting this commit and removing added modules.