"Refactor code to use pointers for discord.Client in several modules"
This commit modifies several modules across the application, including lmnft, premint, exchangeArt, etherscan, and discord modules where discord.Client is now accessed via a pointer.

By using a pointer as opposed to a value directly, we avoid creating a copy when passing the discord.Client instance around. This improves the performance of the application and it aligns the package code with Go standard coding practices which suggest using pointers for mutable data and controller entities.

In addition, some spaces have been added or removed between lines to improve code formatting and hence, readability.